Move types.NewWebSessionRequest to lib/auth/sessions.go.
Move related methods to lib/auth/sessions.go.
